Updates for pep8 compliance.
Some of the long assert lines required more than just whitespace
changes.  In one case, ensuring that -p and -c aren't used at the same
time is now done using the argparse feature for that, replacing the
assert.  In other cases, a new function error_and_exit was added that
prints the arguments of the error message to stderr, then exits with a
non-zero return value.  Then, the asserts are replaced with an if
statement and a call to error_and_exit.